Revert "PDF y funciones mal hechas (no funcionan, revisar)"
This reverts commit c09f6735adb6dec3fc2e91d12065cb9e33026927, reversing
changes made to 094b05bb9903bfc37fbdc90a9fffc324e4722585.

+++ src/main/java/com/tfg/springmarket/model/entities/ProductosEstablecimiento.java
@@ -23,7 +23,7 @@ public class ProductosEstablecimiento implements Serializable {
     private Double precioCoste;
 
     @Column(name = "precio_venta")
-    private Double precioVenta;
+    private BigDecimal precioVenta;
 
     @Column(name = "stock")
     private Integer stock;
